test 2 (final, I think)
For some reason, ICAR.jpg doesn’t work.  icar.jpg is fine, so is
ICAR2.jpg; MPR-flex.jpg also has same problem.  Not sure why - but went
with the lower case instead of spending more time on it.  Note: on my
local machine it worked fine, but not on github.